requirements.txt
argon2-cffi==20.1.0
django-cors-headers==3.4.0
django==3.1.0
djangorestframework-simplejwt==4.4.0
djangorestframework==3.11.1
model-bakery==1.1.1
orjson==3.3.1
pytest-cov==2.10.1
pytest-django==3.9.0
pytest==6.0.1
python-decouple==3.3
requests==2.24.0
wheel==0.35.1